ping_pong_comparison/requirements.txt
websockets
fastapi
uvicorn[standard] # Untuk FastAPI server
requests          # Untuk REST client
grpcio
grpcio-tools
matplotlib
numpy             # matplotlib biasanya butuh ini